+++ NEWS.md
@@ -1,6 +1,6 @@
 version 1.2.3
   - low resolution shapefiles added for republika, kraje and okresy
-  - internet connection not required for using low resolution shapefiles
+  - internet connection not required for using lo-res shapefiles
   - dependency on tidverse replaced by dplyr
 
 version 1.1.1